0.1.3 231 266 267 TX ACK msg
Add validation of RM_MSG_PUSH and do validation
in rm_do_msg_push_rx in receiver.
Send ACk for RM_MSG_PUSH in rm_do_msg_push_rx
in receiver.

+++ src/rm_core.c
@@ -1,8 +1,8 @@
-/* @file	    rm_core.c
- * @brief	    Daemon's start up.
- * @author	    Piotr Gregor <piotrgregor@rsyncme.org>
- * @date	    02 Jan 2016 02:50 PM
- * @copyright	LGPLv2.1 */
+/* @file        rm_core.c
+ * @brief       Daemon's start up.
+ * @author      Piotr Gregor <piotrgregor@rsyncme.org>
+ * @date        02 Jan 2016 02:50 PM
+ * @copyright   LGPLv2.1 */
 
 
 #include "rm_core.h"
@@ -72,7 +72,7 @@ rm_core_session_add(struct rsyncme *rm, struct rm_session *s) {
     pthread_mutex_lock(&rm->mutex);
     twlist_add(&rm->sessions_list, &s->link);
 
-    twhash_add(rm->sessions, &s->hlink, (uint32_t)s->hash.data);
+    twhash_add(rm->sessions, &s->hlink, (uint64_t)s->hash.data);
     rm->sessions_n++;
     pthread_mutex_unlock(&rm->mutex);
     return;
